requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
click==8.1.3
coverage==7.1.0
decorator==5.1.1
executing==1.2.0
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
platformdirs==3.0.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
Pygments==2.14.0
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
